Move ,  and  logic into their cobra commands. For engine and proxy packages, move  functions into utils_ file.
