--- engine/api.go
@@ -1,81 +1,0 @@
-package engine
-
-import (
-	"fmt"
-	"path/filepath"
-	"strings"
-)
-
-// Init - Create a new Konnect object.
-func Init(filename string) (*Konnect, error) {
-	filename, err := filepath.Abs(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	konnect := New()
-	if err = konnect.LoadFromFile(filename); err != nil {
-		return nil, err
-	}
-
-	return konnect, nil
-}
-
-// List - Show info for all SSHProxy objects.
-func (k *Konnect) List() string {
-	result := ""
-	hosts := k.GetHosts()
-	for _, host := range hosts {
-		result += fmt.Sprintln(k.Hosts[host].Info())
-	}
-	return result
-}
-
-// Args - Print SSH Args for a given host.
-func (k *Konnect) Args(host string) (string, error) {
-	proxy, err := k.Get(host)
-	if err != nil {
-		return "", err
-	}
-
-	argsStr := strings.Join(proxy.Args(), " ")
-	return argsStr, err
-}
-
-// Connect to host.
-func (k *Konnect) Connect(host string) error {
-	proxy, err := k.Get(host)
-	if err != nil {
-		return err
-	}
-
-	return proxy.Connect()
-}
-
-// Status - Check the status of one or more hosts.
-func (k *Konnect) Status(hosts []string) {
-	// For each specified host, launch a goroutine
-	// to test the ssh connection.
-	for _, host := range hosts {
-		go k.TestConnection(host)
-	}
-
-	// This goroutine blocks until all specified
-	// host connections have been tested.
-	go func() {
-		for i := 0; i < len(hosts); i++ {
-			<-k.ProxyChan
-		}
-		k.CompletedChan <- true
-	}()
-
-	// Block until above goroutine completes.
-	<-k.CompletedChan
-	// Done
-
-	// Print the status of all hosts.
-	for _, host := range hosts {
-		proxy := k.Hosts[host]
-		proxy.PrintStatus()
-	}
-}
