+++ engine/engine.go
@@ -85,11 +85,30 @@ func (k *Konnect) TestConnection(host string) {
 	k.ProxyChan <- true
 }
 
-// New - Create a new Konnect object.
-func New() *Konnect {
-	return &Konnect{
-		Hosts:         make(map[string]*proxy.SSHProxy),
-		ProxyChan:     make(chan bool),
-		CompletedChan: make(chan bool),
+// Status - Check the status of one or more hosts.
+func (k *Konnect) Status(hosts []string) {
+	// For each specified host, launch a goroutine
+	// to test the ssh connection.
+	for _, host := range hosts {
+		go k.TestConnection(host)
+	}
+
+	// This goroutine blocks until all specified
+	// host connections have been tested.
+	go func() {
+		for i := 0; i < len(hosts); i++ {
+			<-k.ProxyChan
+		}
+		k.CompletedChan <- true
+	}()
+
+	// Block until above goroutine completes.
+	<-k.CompletedChan
+	// Done
+
+	// Print the status of all hosts.
+	for _, host := range hosts {
+		proxy := k.Hosts[host]
+		proxy.PrintStatus()
 	}
 }
