+++ proxy/utils_.go
@@ -21,3 +21,23 @@ func getDefaultKey() string {
 func getDefaultPort() int {
 	return 22
 }
+
+// New - Create a new SSHProxy object with given values.
+func New(user, host string, port int, key string) *SSHProxy {
+	return &SSHProxy{
+		User: user,
+		Host: host,
+		Port: port,
+		Key:  key,
+	}
+}
+
+// Default - Create a new SSHProxy object with default values.
+func Default() *SSHProxy {
+	return &SSHProxy{
+		User: "",
+		Host: "",
+		Port: getDefaultPort(),
+		Key:  getDefaultKey(),
+	}
+}
